Fixed Remove-VMDvdDrive errors while running SCSI tests

+++ WS2012R2/lisa/setupscripts/AddHardDisk.ps1
@@ -122,40 +122,6 @@ param([string] $vmName, [string] $hvServer, [string] $testParams)
 
 $global:MinDiskSize = "1GB"
 
-
-#######################################################################
-#
-# GetRemoteFileInfo()
-#
-# Description:
-#     Use WMI to retrieve file information for a file residing on the
-#     Hyper-V server.
-#
-# Return:
-#     A FileInfo structure if the file exists, null otherwise.
-#
-#######################################################################
-function GetRemoteFileInfo([String] $filename, [String] $server )
-{
-    $fileInfo = $null
-
-    if (-not $filename)
-    {
-        return $null
-    }
-
-    if (-not $server)
-    {
-        return $null
-    }
-
-    $remoteFilename = $filename.Replace("\", "\\")
-    $fileInfo = Get-WmiObject -query "SELECT * FROM CIM_DataFile WHERE Name='${remoteFilename}'" -ComputerName $server
-
-    return $fileInfo
-}
-
-
 ############################################################################
 #
 # CreateController
@@ -223,44 +189,6 @@ function CreateController([string] $vmName, [string] $server, [string] $controll
     }
 }
 
-function ConvertStringToUInt64([string] $newSize)
-{
-    $uint64Size = $null
-    #
-    # Make sure we received a string to convert
-    #
-    if (-not $newSize)
-    {
-        Write-Error -Message "ConvertStringToUInt64() - input string is null" -Category InvalidArgument -ErrorAction SilentlyContinue
-        return $null
-    }
-
-    if ($newSize.EndsWith("MB"))
-    {
-        $num = $newSize.Replace("MB","")
-        $uint64Size = ([Convert]::ToUInt64($num)) * 1MB
-    }
-    elseif ($newSize.EndsWith("GB"))
-    {
-        $num = $newSize.Replace("GB","")
-        $uint64Size = ([Convert]::ToUInt64($num)) * 1GB
-    }
-    elseif ($newSize.EndsWith("TB"))
-    {
-        $num = $newSize.Replace("TB","")
-        $uint64Size = ([Convert]::ToUInt64($num)) * 1TB
-    }
-    else
-    {
-        Write-Error -Message "Invalid newSize parameter: ${newSize}" -Category InvalidArgument -ErrorAction SilentlyContinue
-        return $null
-    }
-
-
-    return $uint64Size
-}
-
-
 ############################################################################
 #
 # CreatePassThruDrive
@@ -317,12 +245,16 @@ function CreatePassThruDrive([string] $vmName, [string] $server, [switch] $scsi,
         }
     }
 
-    $dvd = Get-VMDvdDrive -VMName $vmName -ComputerName $server
-    if ($dvd)
+    $vmGen = GetVMGeneration $vmName $hvServer
+
+    if (($vmGen -eq 1) -and ($controllerType -eq "IDE"))
     {
-        Remove-VMDvdDrive $dvd
+        $dvd = Get-VMDvdDrive -VMName $vmName -ComputerName $server
+        if ($dvd)
+        {
+            Remove-VMDvdDrive $dvd
+        }
     }
-    
 
     # Create the .vhd file if it does not already exist, then create the drive and mount the .vhdx
     $hostInfo = Get-VMHost -ComputerName $server
@@ -443,11 +375,17 @@ function CreateHardDrive( [string] $vmName, [string] $server, [System.Boolean] $
         }
     }
 
-    $dvd = Get-VMDvdDrive -VMName $vmName -ComputerName $hvServer
-    if ($dvd)
+    $vmGen = GetVMGeneration $vmName $hvServer
+
+    if (($vmGen -eq 1) -and ($controllerType -eq "IDE"))
     {
-        Remove-VMDvdDrive $dvd
+        $dvd = Get-VMDvdDrive -VMName $vmName -ComputerName $hvServer
+        if ($dvd)
+        {
+            Remove-VMDvdDrive $dvd
+        }
     }
+
     #
     # Create the .vhd file if it does not already exist
     #
@@ -604,6 +542,11 @@ if ($testParams -eq $null -or $testParams.Length -lt 3)
     return $false
 }
 
+# Source TCUtils
+. .\setupScripts\TCUtils.ps1
+
+# Source STOR_VHDXResize_Utils
+. .\setupScripts\STOR_VHDXResize_Utils.ps1
 
 #
 # Parse the testParams string
